+++ src/const.js
@@ -76,4 +76,12 @@ const SORT_TYPE = {
   PRICE: `sort-price`,
 };
 
-export {TRANSFER_TYPE, ACTIVITY_TYPE, EVENT_DESTINATION, MONTH_NAMES, RANDOM_PRICE_SETTINGS, RENDER_POSITION, SORT_TYPE};
+export {
+  TRANSFER_TYPE,
+  ACTIVITY_TYPE,
+  EVENT_DESTINATION,
+  MONTH_NAMES,
+  RANDOM_PRICE_SETTINGS,
+  RENDER_POSITION,
+  SORT_TYPE
+};
