+++ src/mock-data/trip-event-item-data.js
@@ -1,5 +1,6 @@
-import {getRandomItemFromArray, getRandomNumberFromInterval} from '../utils/common';
 import {TRANSFER_TYPE, ACTIVITY_TYPE, EVENT_DESTINATION, RANDOM_PRICE_SETTINGS} from '../const';
+
+import {getRandomItemFromArray, getRandomNumberFromInterval} from '../utils/common';
 import {generateTripEventOfferData} from './trip-event-offer-data';
 import {generateTripEventDateData} from "./trip-event-date-data";
 import {generateTripEventDestinationData} from "./trip-event-destination-data";
