+++ src/utils/common.js
@@ -1,5 +1,3 @@
-import moment from 'moment';
-
 const getRandomItemFromArray = (arr) => arr[Math.floor(Math.random() * arr.length)];
 
 const getRandomItemsFromArray = (arr, quantity) => {
@@ -15,25 +13,6 @@ const getRandomInt = (max, min = 1) => Math.floor(Math.random() * (max - min)) +
 const removeDuplicatesFromArray = (arr) => Array.from(new Set(arr));
 
 const getRandomNumberFromInterval = (min, max, mult) => Math.floor(Math.floor(Math.random() * (max - min + 1) + min) / mult) * mult;
-
-const castTimeFormat = (value) => value < 10 ? `0${value}` : String(value);
-
-const formatTime = (date) => moment(date).format(`hh:mm`);
-
-const formatDate = (date) => moment(date).format(`DD MMMM`);
-
-const timeDuration = (start, end) => {
-  const momentDiff = moment(end).diff(moment(start));
-  const momentDuration = moment.duration(momentDiff);
-
-  const duration = {
-    days: momentDuration.get(`days`) > 0 ? `${castTimeFormat(momentDuration.get(`days`))}D` : ``,
-    hours: momentDuration.get(`hours`) > 0 ? `${castTimeFormat(momentDuration.get(`hours`))}H` : ``,
-    minutes: momentDuration.get(`minutes`) > 0 ? `${castTimeFormat(momentDuration.get(`minutes`))}M` : ``,
-  };
-
-  return `${duration.days} ${duration.hours} ${duration.minutes}`;
-};
 
 const checkEventType = (type, arr) => {
   const isActivityType = arr.some((item) => item === type);
@@ -62,10 +41,6 @@ export {
   getRandomInt,
   removeDuplicatesFromArray,
   getRandomNumberFromInterval,
-  castTimeFormat,
-  formatTime,
-  timeDuration,
   checkEventType,
   updateItem,
-  formatDate,
 };
